+++ SALESANDINVENTORYSYSTEM/ucPUTINTOINVENTORY.vb
@@ -215,6 +215,11 @@ Public Class ucPUTINTOINVENTORY
     End Sub
 
     Private Sub BTNDELETE_Click_1(sender As Object, e As EventArgs) Handles BTNDELETE.Click
+
+        If MessageBox.Show("Are you sure to remove this from the Stock List?", "Delete", MessageBoxButtons.YesNo, MessageBoxIcon.Information, MessageBoxDefaultButton.Button2) = DialogResult.No Then
+            Exit Sub
+        End If
+
         For i As Integer = 0 To DGVMAIN.SelectedRows.Count - 1
             Dim cmd As New MySqlCommand("delete from tbl_stocks where stockid = @si ", con)
             cmd.Parameters.AddWithValue("si", DGVMAIN.SelectedRows(i).Cells(0).Value.ToString())
@@ -233,25 +238,25 @@ Public Class ucPUTINTOINVENTORY
 
         ' ------------------------------------------------------------- ERROR TRAPPING -------------------------------------------------------------------
 
-        If TXTPI.Text <> TXTSAM.Text Then
-            OpenCon()
-            cmd.CommandText = "Select * from tbl_stocks where prodid = '" & TXTPI.Text & "'"
-            dr = cmd.ExecuteReader()
-            If dr.HasRows Then
-                MsgBox("Product is already added on the stocks", vbOKOnly + vbCritical, "Error Saving")
-                con.Close()
-                TXTPI.Text = ""
-                TXTPNA.Text = ""
-                TXTUNIT.Text = ""
-                CBOUNIT.Text = ""
-                TXTPRICE.Text = ""
-                NUDQUANTITY.Value = 0
-                txtmix.Text = ""
+        'If TXTPI.Text <> TXTSAM.Text Then
+        '    OpenCon()
+        '    cmd.CommandText = "Select * from tbl_stocks where prodid = '" & TXTPI.Text & "'"
+        '    dr = cmd.ExecuteReader()
+        '    If dr.HasRows Then
+        '        MsgBox("Product is already added on the stocks", vbOKOnly + vbCritical, "Error Saving")
+        '        con.Close()
+        '        TXTPI.Text = ""
+        '        TXTPNA.Text = ""
+        '        TXTUNIT.Text = ""
+        '        CBOUNIT.Text = ""
+        '        TXTPRICE.Text = ""
+        '        NUDQUANTITY.Value = 0
+        '        txtmix.Text = ""
 
-                Exit Sub
-            End If
-            con.Close()
-        End If
+        '        Exit Sub
+        '    End If
+        '    con.Close()
+        'End If
 
         If NUDQUANTITY.Text = "" Or TXTPRICE.Text = "" Or TXTUNIT.Text = "" Or CBOUNIT.Text = "" Then
             MsgBox("All fields are required!", vbOKOnly + vbCritical, "Notice")
